Rename FixedPoolSubmitter
